requirements.txt
gspread
telegram-python-bot
